rename test classes
