+++ tests/tests-integrations/src/main/java/org/eclipse/dirigible/integration/tests/TransactionsTestSuite.java
@@ -2,18 +2,18 @@ package org.eclipse.dirigible.integration.tests;
 
 import org.eclipse.dirigible.integration.tests.ui.tests.GitPerspectiveIT;
 import org.eclipse.dirigible.integration.tests.ui.tests.MultitenancyIT;
-import org.eclipse.dirigible.integration.tests.ui.tests.QuartzTransactionsIT;
-import org.eclipse.dirigible.integration.tests.ui.tests.camel.CamelTransactionsIT;
+import org.eclipse.dirigible.integration.tests.ui.tests.QuartzTransactionsRollbackIT;
+import org.eclipse.dirigible.integration.tests.ui.tests.camel.CamelTransactionsRollbackIT;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
 
 @Suite
 @SelectClasses({//
         GitPerspectiveIT.class, //
-        QuartzTransactionsIT.class, //
-        CamelTransactionsIT.class, //
+        QuartzTransactionsRollbackIT.class, //
+        CamelTransactionsRollbackIT.class, //
         MultitenancyIT.class, //
-// RestTransactionsIT.class, //
+        // RestTransactionsIT.class, //
 })
 public class TransactionsTestSuite {
     // use this suite class to run all transaction related tests in the IDE
