+++ tests/tests-integrations/src/main/resources/CamelTransactionsRollbackIT/camel-handler.ts
