+++ tests/tests-integrations/src/main/resources/CamelTransactionsRollbackIT/edm.edm
