+++ tests/tests-integrations/src/main/resources/CamelTransactionsRollbackIT/project.json
