+++ tests/tests-integrations/src/main/resources/CamelTransactionsRollbackIT/test-camel.camel
