+++ tests/tests-integrations/src/main/resources/CamelTransactionsRollbackIT/tsconfig.json
