+++ tests/tests-integrations/src/main/resources/QuartzTransactionsRollbackIT/edm.gen
