+++ tests/tests-integrations/src/main/resources/QuartzTransactionsRollbackIT/jobs/test-job-handler.ts
