+++ tests/tests-integrations/src/main/resources/QuartzTransactionsRollbackIT/project.json
