+++ tests/tests-integrations/src/main/resources/QuartzTransactionsRollbackIT/tsconfig.json
